pyproject.toml
[tool.poetry]
name = "pyparsy"
version = "0.1.0"
description = "HTML parsing library using YAML definitions and XPath"
authors = ["Vladimir Kolev <vladimir.r.kolev@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/vkolev/parsy"
repository = "https://github.com/vkolev/parsy"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Utilities"
]

include = [
    "LICENSE"
]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
lxml = "^4.9.1"
schema = "^0.7.5"
parsel = "^1.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
